+++ Changelog.markdown
@@ -34,6 +34,7 @@ This version has some small incompatibilities with `0.4.1` so read Changelog car
 * __new__:  `diaspy.people.User._fetchstream()` method,
 * __new__:  `diaspy.people.Me()` object representing current user,
 * __new__:  `**kwargs` added to `diaspy.streams.Generic.json()` methdo to give developers control over the creation of JSON,
+* __new__:  `.getHCard()` method added to `diaspy.people.User()`,
 
 
 * __upd__:  `diaspy.connection.Connection.login()` modifies connection object in-place **and** returns it (this allows more fluent API),
@@ -41,10 +42,12 @@ This version has some small incompatibilities with `0.4.1` so read Changelog car
 * __upd__:  `diaspy.connection.Connection._login()` no longer returns status code (if login was unsuccessful it'll raise an exception),
 * __upd__:  better error message in `diaspy.models.Post().__init__()`,
 * __upd__:  `data` variable in `diaspy.models.Post()` renamed to `_data` to indicate that it's considered private,
+* __upd__:  after deleting a post `Activity` stream is purged instead of being refilled (this preserves state of stream which is not reset to last 15 posts),
+* __upd__:  `filterByIDs()` method in `Aspects` stream renamed to `filter()`,
 
 
-* __rem__:  `diaspy.connection.Connection.getUserInfo()` moved to `diaspy.people.Me.getInfo()`,
-* __rem__:  `diaspy.people.Me.getInfo()` (moved from `diaspy.connection.Connection.getUserInfo()`),
+* __rem__:  `diaspy.connection.Connection.getUserInfo()` moved to `diaspy.connection.Connection.getUserData()`,
+* __rem__:  `fetch` parameter removed from `diaspy.connection.Connection.getUserData()`,
 
 
 * __dep__:  `max_time` parameter in `diaspy.streams.*.more()` method is deprecated,